# postman/swagger.yml
openapi: 3.0.0
info:
  title: Be-Endless API (Clean)
  description: >-
    Readable, accurate API collection for Be-Endless backend. Matches current
    routes and request bodies.
  version: 1.0.0
servers:
  - url: http://{{base_url}}
...
paths:
...
  /admin/low-stock:
    get:
      tags:
        - Admin - Inventory
      summary: Low Stock - List
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: threshold
          in: query
          schema:
            type: integer
          example: '10'
        - name: skip
          in: query
          schema:
            type: integer
          example: '0'
        - name: limit
          in: query
          schema:
            type: integer
          example: '20'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /product/{product_id}:
    get:
      tags:
        - Catalog (Public - No Auth Required)
      summary: Product - Get by ID
      parameters:
        - name: product_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /get-products:
    post:
      tags:
        - Catalog (Public - No Auth Required)
      summary: Products - List (filters)
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"categoryId\": \"{{category_id}}\",\n  \"skip\": 0,\n 
                \"limit\": 10,\n  \"minPrice\": 1000,\n  \"maxPrice\": 5000,\n 
                \"sortBy\": \"price\",\n  \"sortOrder\": \"asc\",\n 
                \"isFeatured\": true\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /get-Allproducts:
    post:
      tags:
        - Catalog (Public - No Auth Required)
      summary: Products - All (simple list)
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: '"{\n  \"skip\": 0,\n  \"limit\": 20,\n  \"isActive\": true\n}"'
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /search-products:
    post:
      tags:
        - Catalog (Public - No Auth Required)
      summary: Products - Search
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"searchTerm\": \"wheat\",\n  \"skip\": 0,\n  \"limit\":
                10\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /products/nearby:
    post:
      tags:
        - Catalog (Public - No Auth Required)
      summary: Products - Nearby (requires auth)
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"latitude\": 19.0760,\n  \"longitude\": 72.8777,\n 
                \"maxDistance\": 50000,\n  \"skip\": 0,\n  \"limit\": 10\n}"
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /create-product:
    post:
      tags:
        - Admin - Products
      summary: Admin Products - Create (multipart)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Organic Wheat
                description:
                  type: string
                  example: Premium organic wheat
                price:
                  type: integer
                  example: '2500'
                stockQuantity:
                  type: integer
                  example: '100'
                unit:
                  type: string
                  example: kg
                categoryId:
                  type: string
                  example: '{{category_id}}'
                isFeatured:
                  type: boolean
                  example: 'true'
                isActive:
                  type: boolean
                  example: 'true'
                image:
                  type: string
                  format: binary
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /update-product:
    put:
      tags:
        - Admin - Products
      summary: Admin Products - Update
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"productId\": \"{{product_id}}\",\n  \"price\": 2600,\n 
                \"stockQuantity\": 120,\n  \"isFeatured\": false,\n 
                \"isActive\": true\n}"
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /delete-product:
    delete:
      tags:
        - Admin - Products
      summary: Admin Products - Delete
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /update-stock:
    patch:
      tags:
        - Admin - Products
      summary: Admin Products - Update Stock
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"productId\": \"{{product_id}}\",\n  \"stockQuantity\":
                200,\n  \"operation\": \"set\"\n}"
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /toggle-status:
    patch:
      tags:
        - Admin - Products
      summary: Admin Products - Toggle Status
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: '"{\n  \"productId\": \"{{product_id}}\"\n}"'
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /toggle-featured:
    patch:
      tags:
        - Admin - Products
      summary: Admin Products - Toggle Featured
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: '"{\n  \"productId\": \"{{product_id}}\"\n}"'
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{access_token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...